+++ aiter/ops/triton/_triton_kernels/quant/fused_mxfp4_quant.py
@@ -8,7 +8,8 @@ from .quant import _mxfp4_quant_op
 def _rmsmorm_op(row, weight, n_cols, epsilon):
     row_norm = row * row
     row_norm = tl.sum(row_norm, axis=-1)
-    norm_factor = tl.math.rsqrt((row_norm / n_cols) + epsilon)
+    # Cast to f32 before rsqrt to prevent f64 promotion from Python float epsilon.
+    norm_factor = tl.math.rsqrt(((row_norm / n_cols) + epsilon).to(tl.float32))
 
     rms_norm = row * norm_factor[:, None] * weight
     return rms_norm
